+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+## 0.2.1
+- Fixed a bug where an empty string could be used as a cache key.
+- For compatibility, get, contains, and remove will log a warning instead of throwing an error when given an empty key.
+
 ## 0.2.0
 - Added `readAll` method to the backend for retrieving all cache entries.
-
 
 ## 0.1.0